Transfer iTunes test to iTunesProvider
